Update kepler_orbit_final.py
Include Black-Body Spectrum and HR spectrum separely as well as a research tab for the desired star, where the user can search a star, click on the desired result and if all information is there, the star animation will be plotted. Also, there are better error handling pop-ups for the plotting of the graphs and animations.